Replace f32 usage with Decimal
Decimal is now used to retain decimal accuracy. As per the requirements of the OCPP specification.

+++ src/v1_6/types/charging_schedule.rs
@@ -1,4 +1,5 @@
 use chrono::{DateTime, Utc};
+use rust_decimal::Decimal;
 
 use super::{ChargingRateUnitType, ChargingSchedulePeriod};
 
@@ -18,5 +19,6 @@ pub struct ChargingSchedule {
     pub charging_schedule_period: Vec<ChargingSchedulePeriod>,
     /// Optional. Minimum charging rate supported by the electric vehicle. The unit of measure is defined by the chargingRateUnit. This parameter is intended to be used by a local smart charging algorithm to optimize the power allocation for in the case a charging process is inefficient at lower charging rates. Accepts at most one digit fraction (e.g. 8.1)
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub min_charging_rate: Option<f32>,
+    #[serde(with = "rust_decimal::serde::arbitrary_precision_option")]
+    pub min_charging_rate: Option<Decimal>,
 }
