+++ src/v2_0_1/datatypes/charging_schedule_period_type.rs
@@ -1,3 +1,5 @@
+use rust_decimal::Decimal;
+
 /// Charging schedule period structure defines a time period in a charging schedule
 /// ChargingSchedulePeriodType is used by: Common:ChargingScheduleType , Common:CompositeScheduleType
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
@@ -6,7 +8,8 @@ pub struct ChargingSchedulePeriodType {
     /// Required. Start of the period, in seconds from the start of schedule. The value of StartPeriod also defines the stop time of the previous period.
     pub start_period: i32,
     /// Required. Charging rate limit during the schedule period, in the applicable chargingRateUnit, for example in Amperes (A) or Watts (W). Accepts at most one digit fraction (e.g. 8.1).
-    pub limit: f32,
+    #[serde(with = "rust_decimal::serde::arbitrary_precision")]
+    pub limit: Decimal,
     /// Optional. The number of phases that can be used for charging. If a number of phases is needed, numberPhases=3 will be assumed unless another number is given.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub number_phases: Option<i32>,
