+++ src/v2_0_1/messages/cost_updated.rs
@@ -1,4 +1,5 @@
 //! CostUpdated
+use rust_decimal::Decimal;
 use validator::Validate;
 
 /// CostUpdatedRequest, sent by the CSMS to the Charging Station.
@@ -8,7 +9,8 @@ use validator::Validate;
 #[serde(rename_all = "camelCase")]
 pub struct CostUpdatedRequest {
     /// Current total cost, based on the information known by the CSMS, of the transaction including taxes. In the currency configured with the configuration Variable: [Currency]
-    pub total_cost: f32,
+    #[serde(with = "rust_decimal::serde::arbitrary_precision")]
+    pub total_cost: Decimal,
     /// Transaction Id of the transaction the current cost are asked for.
     #[validate(length(min = 0, max = 36))]
     pub transaction_id: String,
